+++ ICD.Connect.Routing.CrestronPro/Receivers/AbstractEndpointReceiverBaseAdapter.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using ICD.Connect.Misc.CrestronPro.Devices;
 using ICD.Connect.Routing.Controls;
 using ICD.Connect.Routing.CrestronPro.Cards;
 using ICD.Connect.Settings;
@@ -221,6 +222,17 @@ namespace ICD.Connect.Routing.CrestronPro.Receivers
 		public DigitalInput GetDigitalInputPort(int address)
 		{
 			string message = string.Format("{0} has no {1}", this, typeof(DigitalInput).Name);
+			throw new ArgumentOutOfRangeException("address", message);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public virtual Cec GetCecPort(eInputOuptut io, int address)
+		{
+			string message = string.Format("{0} has no {1}", this, typeof(Cec).Name);
 			throw new ArgumentOutOfRangeException("address", message);
 		}
 #endif