+++ ICD.Connect.Routing.CrestronPro/Receivers/DmRmc200S/DmRmc200SAdapter.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using ICD.Connect.Misc.CrestronPro.Devices;
 #if SIMPLSHARP
 using Crestron.SimplSharpPro;
 using Crestron.SimplSharpPro.DM;
@@ -77,6 +78,24 @@ namespace ICD.Connect.Routing.CrestronPro.Receivers.DmRmc200S
 				return Receiver.IROutputPorts[(uint)address];
 
 			return base.GetIrOutputPort(address);
+		}
+
+		/// <summary>
+		/// Gets the port at the given address.
+		/// </summary>
+		/// <param name="io"></param>
+		/// <param name="address"></param>
+		/// <returns></returns>
+		public override Cec GetCecPort(eInputOuptut io, int address)
+		{
+			if (Receiver == null)
+				throw new InvalidOperationException("No DmRx instantiated");
+
+			if (io == eInputOuptut.Output && address == 1)
+				return Receiver.HdmiOutput.StreamCec;
+
+			string message = string.Format("No CecPort at address {1}:{2} for device {0}", this, io, address);
+			throw new InvalidOperationException(message);
 		}
 #endif