+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - DMC-SO-HD DM 8G Fiber Output Card
  - DM-RMC-200-S DM 8G Fiber Receiver
+ - CEC Port support for all DM devices
 
 ## [14.0.0] - 2019-01-10
 ### Added