+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - DMC-SO-HD DM 8G Fiber Output Card
+ - DM-RMC-200-S DM 8G Fiber Receiver
+ - CEC Port support for all DM devices
 
 ## [14.0.0] - 2019-01-10
 ### Added