+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [13.1.1] - 2019-01-22
 ### Changed
  - Forcing NVX to use auto-initiation, disabling auto routing